+++ tests/testthat/test_group.R
@@ -13,6 +13,7 @@ test_that("Group methods", {
 
     expect_error(group_job(empty_idf(8.8)), class = "error_idf_not_local")
     expect_silent(group_job(path_idfs, path_epws[1L]))
+    expect_silent(group_job(path_idfs[1], path_epws))
     expect_silent(grp <- group_job(path_idfs, NULL))
     expect_equal(grp$status(),
         list(run_before = FALSE, alive = FALSE, terminated = NA,