+++ tests/testthat/test_install.R
@@ -4,4 +4,10 @@ test_that("Install", {
     if (is_avail_eplus(8.8)) expect_error(install_eplus(8.8, local = TRUE))
     if (is_macos()) expect_error(install_eplus(8.8, local = TRUE))
     else install_eplus(8.8, local = TRUE, force = TRUE)
+
+    # test if patch on EnergyPlus v9.1 and above works
+    if (!is_avail_eplus(9.1)) install_eplus(9.1)
+    if (!is_avail_eplus(9.2)) install_eplus(9.2)
+    expect_true(is_avail_eplus(9.1))
+    expect_true(is_avail_eplus(9.2))
 })