+++ NEWS.md
@@ -1,3 +1,5 @@
+# eplusr (development version)
+
 # eplusr 0.11.0
 
 ## New features