+++ NEWS.md
@@ -6,6 +6,11 @@
 * A new method `Idf$last_job()` has been added to enable getting the last
   simulation job created using `Idf$run()` (#187).
 
+## Bug fixes
+
+* The algorithm of detecting numeric columns in `EplusSql$tabular_data()` has
+  been improved (#190).
+
 # eplusr 0.11.0
 
 ## New features