+++ NEWS.md
@@ -31,6 +31,7 @@
 * Now `eplus_config()` will always return the expanded EnergyPlus path (#196).
 * Now `group_job()` will return more informative error messages when input
   contains `Idf` objects that havn't been saved (#204).
+* Fix error in `EplusGroupJob$run()` when custom `dir` is specified (#205).
 
 # eplusr 0.11.0